feat: Add useFetchAlbumsQuery hook for fetching albums
This commit adds the `useFetchAlbumsQuery` hook to the `albumsApi` module. The hook is used to fetch albums for a specific user. It makes a GET request to the `/albums` endpoint with the `userId` as a query parameter. This hook will be used in the `AlbumsList` component to fetch albums for a given user.