feat: Add lazy loading for images in AlbumsList
This commit adds lazy loading functionality for images in the AlbumsList component. The useFetchAlbumsQuery hook has been updated to include the isFetching property, which indicates whether the data is currently being fetched. The handleAddAlbum function has been implemented to add a new album for the user. The loading state has been handled with a Skeleton component while the data is being fetched.